// src/enum/academic-class.enum.ts
export enum AcademicClassType {
  CLASS_I = 'Class I',
  CLASS_II = 'Class II',
  CLASS_III = 'Class III',
  CLASS_IV = 'Class IV',
  CLASS_V = 'Class V',
  CLASS_VI = 'Class VI',
  CLASS_VII = 'Class VII',
  CLASS_VIII = 'Class VIII',
  CLASS_IX = 'Class IX',
  CLASS_X = 'Class X',
  CLASS_XI = 'Class XI',
  CLASS_XII = 'Class XII',
}